azure-functions/pyproject.toml
[project]
name = "office-addins-mcp-server-azure"
version = "0.1.0"
description = "Azure Functions MCP server for discovering and managing Microsoft Office Add-ins across Word, Excel, PowerPoint, Outlook, and Teams"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    { name = "Yi-Wei Chen", email = "feberium@gmail.com" }
]
keywords = ["mcp", "office", "addins", "microsoft", "azure-functions", "serverless"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Office/Business",
    "Operating System :: OS Independent",
    "Framework :: AsyncIO",
    "Environment :: Web Environment",
]
dependencies = [
    "azure-functions>=1.19.0",
    "azure-functions-worker>=1.19.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.metadata]
allow-direct-references = true

# Azure Functions specific configuration
[tool.azure-functions]
project-type = "python"
python-version = "3.11"
runtime-version = "~4"